wip consolidate table row components
