--- app/templates/components/demographic-change-table-row.hbs
@@ -1,225 +1,0 @@
-{{#if subheader}}
-
-  <td colspan="6">{{title}}</td>
-  <td>&nbsp;</td>
-  {{#if reliability}}
-    <td>&nbsp;</td>
-    <td>&nbsp;</td>
-  {{/if}}
-  {{#if comparison}}
-    <td>&nbsp;</td>
-    <td>&nbsp;</td>
-    <td>&nbsp;</td>
-    <td>&nbsp;</td>
-    {{#if reliability}}
-      <td>&nbsp;</td>
-      <td>&nbsp;</td>
-      <td>&nbsp;</td>
-      <td>&nbsp;</td>
-    {{/if}}
-  {{/if}}
-
-{{else}}{{! subheader = false }}
-
-{{#if divider}}
-
-  <td>&nbsp;</td>
-  <td class="cell-border-left">&nbsp;</td>
-  <td>&nbsp;</td>
-  <td class="cell-border-left">&nbsp;</td>
-  <td>&nbsp;</td>
-  <td class="cell-border-left">&nbsp;</td>
-  {{#if reliability}}
-    <td>&nbsp;</td>
-  {{/if}}
-  <td>&nbsp;</td>
-  {{#if reliability}}
-    <td>&nbsp;</td>
-  {{/if}}
-  {{#if comparison}}
-    <td class="cell-border-left">&nbsp;</td>
-    {{#if reliability}}
-      <td>&nbsp;</td>
-    {{/if}}
-    <td>&nbsp;</td>
-    {{#if reliability}}
-      <td>&nbsp;</td>
-    {{/if}}
-    <td class="cell-border-left">&nbsp;</td>
-    {{#if reliability}}
-      <td>&nbsp;</td>
-    {{/if}}
-    <td>&nbsp;</td>
-    {{#if reliability}}
-      <td>&nbsp;</td>
-    {{/if}}
-  {{/if}}
-
-{{else}}{{! subheader && divider = false }}
-
-  <td>{{title}}</td>
-  <td class="cell-border-left">
-    {{#if data2.sum}}
-      {{numeral-format data2.sum}}
-    {{else if (eq data2.sum 0)}}
-      {{numeral-format data2.sum}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  <td>
-    {{#if data2.percent}}
-      {{numeral-format data2.percent '0.0%'}}
-    {{else if (eq data2.percent 0)}}
-      {{numeral-format data2.percent '0.0%'}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  <td class="cell-border-left">
-    {{#if data.sum}}
-      {{numeral-format data.sum}}
-    {{else if (eq data.sum 0)}}
-      {{numeral-format data.sum}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  <td>
-    {{#if data.percent}}
-      {{numeral-format data.percent '0.0%'}}
-    {{else if (eq data.percent 0)}}
-      {{numeral-format data.percent '0.0%'}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  <td class="cell-border-left">
-    {{#if data2.sum}}
-      {{numeral-format (sub data.sum data2.sum) '+0,0'}}
-    {{else if (eq data2.sum 0)}}
-      {{numeral-format (sub data.sum data2.sum) '+0,0'}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  {{#if reliability}}
-  <td>
-    {{#if data2.sum}}
-      {{numeral-format (sqrt (add (pow (div data2.m 1.645) 2) (pow (div data.m 1.645) 2))) '+0,0'}}
-    {{else if (eq data2.sum 0)}}
-      {{numeral-format (sqrt (add (pow (div data2.m 1.645) 2) (pow (div data.m 1.645) 2))) '+0,0'}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  {{/if}}
-  <td>
-    {{#if data2.percent}}
-      {{numeral-format (div (sub data.sum data2.sum) data2.sum) '+0.0%'}}
-    {{else if (eq data2.percent 0)}}
-      {{numeral-format (div (sub data.sum data2.sum) data2.sum) '+0.0%'}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  {{#if reliability}}
-  <td>
-    {{#if data2.sum}}
-      {{numeral-format longitudinalPercentMOE '+0.0%'}}
-    {{else if (eq data2.sum 0)}}
-      {{numeral-format longitudinalPercentMOE '+0.0%'}}
-    {{else}}
-      n/a
-    {{/if}}
-
-  </td>
-  {{/if}}
-  {{#if comparison}}
-  <td class="cell-border-left">
-    {{#if data2.comparison_sum}}
-      {{numeral-format (sub data.comparison_sum data2.comparison_sum) '+0,0'}}
-    {{else if (eq data2.comparison_sum 0)}}
-      {{numeral-format (sub data.comparison_sum data2.comparison_sum) '+0,0'}}
-    {{else}}
-    n/a
-    {{/if}}
-  </td>
-  {{#if reliability}}
-  <td>
-    {{#if data2.comparison_sum}}
-      {{numeral-format (sqrt (add (pow (div data2.comparison_m 1.645) 2) (pow (div data.comparison_m 1.645) 2))) '+0,0'}}
-    {{else if (eq data2.comparison_sum 0)}}
-      {{numeral-format (sqrt (add (pow (div data2.comparison_m 1.645) 2) (pow (div data.comparison_m 1.645) 2))) '+0,0'}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  {{/if}}
-  <td>
-    {{#if data2.comparison_percent}}
-      {{numeral-format (div (sub data.comparison_sum data2.comparison_sum) data2.comparison_sum) '+0.0%'}}
-    {{else if (eq data2.comparison_percent 0)}}
-      {{numeral-format (div (sub data.comparison_sum data2.comparison_sum) data2.comparison_sum) '+0.0%'}}
-    {{else}}
-    n/a
-    {{/if}}
-  </td>
-  {{#if reliability}}
-  <td>
-    {{#if data2.sum}}
-      {{numeral-format comparisonLongitudinalPercentMOE '+0.0%'}}
-    {{else if (eq data2.sum 0)}}
-      {{numeral-format comparisonLongitudinalPercentMOE '+0.0%'}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  {{/if}}
-  <td class="cell-border-left">
-    {{#if data2.sum}}
-      {{numeral-format (sub (sub data.sum data2.sum) (sub data.comparison_sum data2.comparison_sum)) '+0,0'}}
-    {{else if (eq data2.sum 0)}}
-      {{numeral-format (sub (sub data.sum data2.sum) (sub data.comparison_sum data2.comparison_sum)) '+0,0'}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  {{#if reliability}}
-    <td>
-    {{#if data2.sum}}
-      {{numeral-format differenceMOE '+0,0'}}
-    {{else if (eq data2.sum 0)}}
-      {{numeral-format differenceMOE '+0,0'}}
-    {{else}}
-      n/a
-    {{/if}}
-    </td>
-  {{/if}}
-  <td>
-    {{#if data2.percent}}
-      {{numeral-format (mult (sub (div (sub data.sum data2.sum) data2.sum) (div (sub data.comparison_sum data2.comparison_sum) data2.comparison_sum)) 100) '+0.0'}}
-    {{else if (eq data2.percent 0)}}
-      {{numeral-format (mult (sub (div (sub data.sum data2.sum) data2.sum) (div (sub data.comparison_sum data2.comparison_sum) data2.comparison_sum)) 100) '+0.0'}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  {{#if reliability}}
-  <td>
-    {{#if data2.sum}}
-      {{numeral-format differencePercentMOE '+0.0%'}}
-    {{else if (eq data2.sum 0)}}
-      {{numeral-format differencePercentMOE '+0.0%'}}
-    {{else}}
-      n/a
-    {{/if}}
-  </td>
-  {{/if}}
-  {{/if}}
-
-  {{/if}}
-
-  {{/if}}
-
-  {{yield}}
