+++ app/templates/components/demographic-table.hbs
@@ -1,39 +1,101 @@
+
+{{log options}}
 <div class="table-scroll">
-  <table class="text-small {{if scrollTable 'table-scroll-vert'}}">
+  <table class="text-small {{if options.scrollTable 'table-scroll-vert'}}">
     <thead>
+      {{#if (eq options.mode 'current')}} {{! current header}}
       <tr>
         <th rowspan="2">&nbsp;</th>
-        <th colspan="{{if reliability "5" "2"}}" class="text-center cell-border-bottom">Selected Area</th>
-        {{#if comparison}}
-        <th colspan="{{if reliability "5" "2"}}" class="text-center cell-border-bottom">Comparison Area</th>
+        <th colspan="{{if options.reliability "5" "2"}}" class="text-center cell-border-bottom">Selected Area</th>
+        {{#if options.comparison}}
+        <th colspan="{{if options.reliability "5" "2"}}" class="text-center cell-border-bottom">Comparison Area</th>
         <th colspan="2" class="text-center cell-border-bottom">Difference</th>
         {{/if}}
       </tr>
       <tr>
         <th class="text-right cell-border-bottom">Estimate</th>
-          {{#if reliability}}
+          {{#if options.reliability}}
           <th class="text-right cell-border-bottom">MOE</th>
         <th class="text-right cell-border-bottom">CV</th>
         {{/if}}
         <th class="text-right cell-border-bottom">Pct. Estimate</th>
-          {{#if reliability}}
+          {{#if options.reliability}}
         <th class="text-right cell-border-bottom">MOE</th>
           {{/if}}
-       {{#if comparison}}
+       {{#if options.comparison}}
           <th class="text-right cell-border-bottom">Estimate</th>
-          {{#if reliability}}
+          {{#if options.reliability}}
         <th class="text-right cell-border-bottom">MOE</th>
         <th class="text-right cell-border-bottom">CV</th>
           {{/if}}
         <th class="text-right cell-border-bottom">Pct. Estimate</th>
-          {{#if reliability}}
+          {{#if options.reliability}}
         <th class="text-right cell-border-bottom">MOE</th>
           {{/if}}
         <th class="text-right cell-border-bottom">Number</th>
         <th class="text-right cell-border-bottom">Pctg. Point</th>
       {{/if}}
       </tr>
+      {{ else }} {{! change over time header}}
+        <tr>
+          <th rowspan="3">&nbsp;</th>
+          <th colspan="2" rowspan="2" class="text-center cell-border-bottom">2006-2010</th>
+          <th colspan="2" rowspan="2" class="text-center cell-border-bottom">2011-2015</th>
+          {{#if options.comparison}}
+            {{#if options.reliability}}
+              <th colspan="12" class="text-center cell-border-bottom">Change, 2006-2010 to 2011-2015</th>
+            {{else}}
+              <th colspan="6" class="text-center cell-border-bottom">Change, 2006-2010 to 2011-2015</th>
+            {{/if}}
+          {{else}}
+            {{#if options.reliability}}
+              <th colspan="4" class="text-center cell-border-bottom">Change, 2006-2010 to 2011-2015</th>
+            {{else}}
+              <th colspan="2" class="text-center cell-border-bottom">Change, 2006-2010 to 2011-2015</th>
+              {{/if}}
+          {{/if}}
+        </tr>
+        <tr>
+          <th colspan="{{if options.reliability "4" "2"}}" class="text-center cell-border-bottom">Selected Area</th>
+          {{#if options.comparison}}
+          <th colspan="{{if options.reliability "4" "2"}}" class="text-center cell-border-bottom">Comparison Area</th>
+          <th colspan="{{if options.reliability "4" "2"}}" class="text-center cell-border-bottom">Difference</th>
+          {{/if}}
+        </tr>
+        <tr>
+          <th>Number</th>
+          <th>Percent</th>
+          <th>Number</th>
+          <th>Percent</th>
+          <th>Estimate</th>
+          {{#if options.reliability}}
+            <th>MOE</th>
+          {{/if}}
+          <th>Pct. Estimate</th>
+          {{#if options.reliability}}
+            <th>MOE</th>
+          {{/if}}
+          {{#if options.comparison}}
+          <th>Estimate</th>
+          {{#if options.reliability}}
+            <th>MOE</th>
+          {{/if}}
+          <th>Pct. Estimate</th>
+          {{#if options.reliability}}
+            <th>MOE</th>
+          {{/if}}
+          <th>Estimate</th>
+          {{#if options.reliability}}
+            <th>MOE</th>
+          {{/if}}
+          <th>Pctg. Estimate</th>
+          {{#if options.reliability}}
+            <th>MOE</th>
+          {{/if}}
+          {{/if}}
+        </tr>
 
+      {{/if}}
     </thead>
     <tbody>
       {{yield}}
